move recorder server into src/recorder
